Change ArbitraryValueToolTip to not jump focus
This changes ArbitraryValueToolTip to use Qt::WA_ShowWithoutActivating,
so that it will show to the screen but not take focus away from the main
window. On Windows, this fixes the cosmetic issue of the main window
losing its activated status (causing the window title bar to change
colors) as the user moves the mouse around and hovers over arbitrary
values.

The ArbitraryValueToolTip can still be interacted with, including
selecting text with the mouse.